Fix for nontermination in GlobalModelCompileErrorsTest.  Utilities.clearEventQueue() is being called and successfully running *from* the event queue, violating the implied semantics and raising the possibility that the method will loop infinitely within the event thread waiting for the queue to empty itself.
The fix for clearEventQueue() is to use the Swing invokeAndWait method which throws an exception if run from the event thread.  This converts the intermittent failure to a hard failure; the *temporary* fix for code that then consistently fails is to comment out calls to clearEventQueue().  However, this may lead to regressions of other intermittent failures, and a better solution is probably needed.

Specific code that needs to be addressed:
ColoringGlyphPainter, line 333
GlobalModelTestCase, line 658

(This commit also contains some cleanup in response to various Eclipse compiler warnings, such as invalid javadoc code and use of raw types.)


git-svn-id: file:///tmp/test-svn/trunk@4941 fe72c1cf-3628-48e9-8b72-1c46755d3cff

+++ drjava/src/edu/rice/cs/util/sexp/NumberAtom.java
@@ -40,7 +40,7 @@ public class NumberAtom implements Atom {
   private double _num;
   private boolean _hasDecimals;
   public NumberAtom(int num){ 
-    _num = (double)num;
+    _num = num;
     _hasDecimals = false;
   }
   public NumberAtom(double num){
